@@ -1,37 +1,26 @@
 {
-  "name": "ts-structure-parser",
-  "version": "0.0.12",
+  "name": "ts-file-parser",
+  "version": "0.0.1",
   "private": false,
   "main": "dist/index.js",
   "scripts": {
-    "build": "rimraf dist && tsc",
-    "pullall" : "dev-env-installer pullall",
-    "buildall" : "dev-env-installer buildall",
-    "testall" : "dev-env-installer testall",
-    "devInstall": "dev-env-installer install"
+    "build": "grunt build",
+    "test": "grunt test"
   },
   "dependencies": {
-    "underscore": "^1.8.3",
-    "typescript": "1.8.7"
-  },
-  "typings": "./dist/index.d.ts",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/mulesoft-labs/ts-structure-parser.git"
-  },
-  "keywords": [
-    "typescript",
-    "content",
-    "parser"
-  ],
-  "homepage": "https://github.com/mulesoft-labs/ts-structure-parser",
-  "license": "Apache-2.0",
-  "bugs": {
-    "url": "https://github.com/mulesoft-labs/ts-structure-parser/issues"
+    "typescript": "^2.5.3"
   },
   "devDependencies": {
-    "typings": "^0.5.1",
-    "dev-env-installer":"0.0.5",
-    "rimraf":"*"
+    "@types/node": "^8.0.31",
+    "@types/chai": "^4.0.4",
+    "@types/mocha": "^2.2.43",
+    "grunt": "^1.0.1",
+    "grunt-ts": "^6.0.0-beta.16",
+    "grunt-tslint": "^5.0.1",
+    "tslint": "^5.7.0",
+    "chai": "^4.1.1",
+    "grunt-mocha-test": "^0.13.2",
+    "mocha": "^3.5.0",
+    "ts-node": "^3.3.0"
   }
 }
