@@ -1,20 +1,19 @@
-/// <reference path="../typings/main.d.ts" />
-import pth=require("path")
-import fs=require("fs")
+import * as pth from "path";
+import * as fs from "fs";
 
 
-export function resolve(p1:string,p2:string):string{
-    return pth.resolve(p1,p2);
+export function resolve(p1: string, p2: string): string {
+    return pth.resolve(p1, p2);
 }
 
-export function readFileSync(p:string):string{
+export function readFileSync(p: string): string {
     return fs.readFileSync(p).toString();
 }
 
-export function dirname(p:string){
+export function dirname(p: string) {
     return pth.dirname(p);
 }
 
-export function existsSync(p:string):boolean{
+export function existsSync(p: string): boolean {
     return fs.existsSync(p);
 }
