@@ -1,29 +1,29 @@
-import index=require("./index");
+import index = require("../index");
 import helperMethodExtractor = require("./helperMethodExtractor");
 
-var ns:{[key:string]:boolean} = {'RamlWrapper':true};
+var ns: {[key: string]: boolean} = {"RamlWrapper": true};
 
-export class HelperMethod{
+export class HelperMethod {
 
     constructor(
-        public originalName:string,
-        public wrapperMethodName:string,
-        public returnType:index.TypeModel,
-        public args:Arg[],
-        public meta:Meta) {
+        public originalName: string,
+        public wrapperMethodName: string,
+        public returnType: index.TypeModel,
+        public args: Arg[],
+        public meta: Meta) {
 
     }
-    targetWrappers():string[]{
+    targetWrappers(): string[] {
 
         var isValid = true;
-        var result:string[] = [];
-        this.args.forEach(x=>{
+        var result: string[] = [];
+        this.args.forEach(x => {
 
-            var arr = flatten(x.type,ns);
-            if(arr.length==0){
+            var arr = flatten(x.type, ns);
+            if (arr.length === 0) {
                 return;
             }
-            if(!isValid || result.length!=0){
+            if (!isValid || result.length !== 0) {
                 result = [];
                 isValid = false;
                 return;
@@ -33,9 +33,9 @@ export class HelperMethod{
         return result;
     }
 
-    callArgs():Arg[]{
-        return this.args.map(x=>{
-            if(flatten(x.type,ns).length==0){
+    callArgs(): Arg[] {
+        return this.args.map(x => {
+            if (flatten(x.type, ns).length === 0) {
                 return x;
             }
             return {
@@ -43,75 +43,71 @@ export class HelperMethod{
                 type: null,
                 optional: false,
                 defaultValue: undefined
-            }
+            };
         });
     }
 }
 
-export interface Arg{
+export interface Arg {
 
-    name:string;
+    name: string;
 
-    type:index.TypeModel;
-    
+    type: index.TypeModel;
+
     defaultValue: any;
-    
+
     optional: boolean;
 }
 
-export interface Meta{
+export interface Meta {
 
-    name?:string
+    name?: string;
 
-    comment?:string
+    comment?: string;
 
-    override?:boolean
+    override?: boolean;
 
-    primary?:boolean
+    primary?: boolean;
 
-    deprecated?: boolean
+    deprecated?: boolean;
 }
 
-export function flatten(t:index.TypeModel, namespaces?:{[key:string]:boolean}):string[]{
+export function flatten(t: index.TypeModel, namespaces?: {[key: string]: boolean}): string[] {
 
-    if(t.typeKind==index.TypeKind.ARRAY){
-        if(namespaces) {
+    if (t.typeKind === index.TypeKind.ARRAY) {
+        if (namespaces) {
             return [];
+        } else {
+            return [ flatten((<index.ArrayType>t).base)[0] + "[]" ];
         }
-        else{
-            return [ flatten((<index.ArrayType>t).base)[0]+'[]' ];
-        }
-    }
-    else if(t.typeKind==index.TypeKind.BASIC){
+    } else if (t.typeKind === index.TypeKind.BASIC) {
         var bt = (<index.BasicType>t);
 
         var str = bt.basicName;
         var nameSpace = bt.nameSpace && bt.nameSpace.trim();
-        if(nameSpace!=null && nameSpace.length>0 && nameSpace!="RamlWrapper"){
+        if (nameSpace && nameSpace.length > 0 && nameSpace !== "RamlWrapper") {
             str = nameSpace + "." + str;
         }
-        if(bt.typeArguments && bt.typeArguments.length!=0){
-            str += `<${bt.typeArguments.map(x=>flatten(x)).join(', ')}>`
+        if (bt.typeArguments && bt.typeArguments.length !== 0) {
+            str += `<${bt.typeArguments.map(x => flatten(x)).join(", ")}>`;
         }
-        if(namespaces) {
+        if (namespaces) {
             if (bt.nameSpace) {
                 return namespaces[bt.nameSpace] ? [ str ] : [];
-            }
-            else{
+            } else {
                 return [];
             }
         }
         return [ str ];
-    }
-    else if (t.typeKind==index.TypeKind.UNION){
+    } else if (t.typeKind === index.TypeKind.UNION) {
         var ut = <index.UnionType>t;
-        var result:string[] = [];
-        ut.options.forEach(x=>result=result.concat(flatten(x,namespaces)));
+        var result: string[] = [];
+        ut.options.forEach(x => result = result.concat(flatten(x, namespaces)));
         return result;
     }
     return [];
 }
 
-export function getHelperMethods(srcPath:string):HelperMethod[]{
+export function getHelperMethods(srcPath: string): HelperMethod[] {
     return helperMethodExtractor.getHelperMethods(srcPath);
 }
